+++ STACpopulator/implementations/DirectoryLoader/crawl_directory.py
@@ -3,6 +3,7 @@ import logging
 import os.path
 import sys
 from typing import Any, MutableMapping, Optional
+import warnings
 
 from requests.sessions import Session
 
@@ -70,10 +71,15 @@ def runner(ns: argparse.Namespace) -> int:
 
 
 def main(*args: str) -> int:
+    warnings.warn(
+        "Calling implementation scripts directly is deprecated. Please use the 'stac-populator' CLI instead.",
+        DeprecationWarning
+    )
     parser = argparse.ArgumentParser()
     add_parser_args(parser)
     ns = parser.parse_args(args or None)
     ns.populator = os.path.basename(os.path.dirname(__file__))
+    ns.command = "run"
     return cli.run(ns)